pyproject.toml
[tool.poetry]
name = "thunderstore-discord-bot"
version = "0.1.0"
description = ""
authors = ["SirQuacker <49299629+SirQuacker@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"
python-dotenv = "^0.15.0"
requests = "^2.25.1"
PyJWT = "^2.0.1"
cryptography = "^3.4.5"
fuzzywuzzy = "^0.18.0"
levenshtein = "^0.12.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
